Add HS2Container
