+++ pom.xml
@@ -47,7 +47,7 @@
     </scm>
 
     <properties>
-        <maven.compiler.release>8</maven.compiler.release>
+        <maven.compiler.release>17</maven.compiler.release>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 
         <apache-hive-jdbc.version>4.0.1</apache-hive-jdbc.version>
